added pv search and lmr

+++ param.h
@@ -11,7 +11,7 @@ constexpr uint8_t EXACT_FLAG = 0;
 constexpr uint8_t ALPHA_FLAG = 1;
 constexpr uint8_t BETA_FLAG = 2;
 
-constexpr uint8_t MAX_DEPTH = 255;
+constexpr int16_t MAX_DEPTH = 255;
 
 constexpr int32_t BASE_SCORE = (1 << 30);
