+++ table.h
@@ -16,7 +16,7 @@ public:
     uint64_t m_hash = 0;
     int32_t m_score = 0;
     chess::Move m_best_move = chess::Move::NULL_MOVE;
-    uint8_t m_depth = 0;
+    int16_t m_depth = 0;
     uint8_t m_flag = 0;
 
     [[nodiscard]] table_entry_result get(uint64_t hash, uint8_t ply, uint8_t depth, int32_t alpha, int32_t beta) const
